fix transaction-logger fix block config add operator key validator
