+++ guardian-service/.env.docker
@@ -10,4 +10,4 @@ OPERATOR_KEY="..."
 INITIALIZATION_TOPIC_ID="0.0.46022543"
 HEDERA_NET="testnet"
 MESSAGE_LANG="en-US"
-LOG_LEVEL="2"
+LOG_LEVEL="1"
